+++ test/components/bullet/bullet.e2e-spec.js
@@ -16,6 +16,38 @@ describe('Bullet example-index tests', () => {
     await utils.checkForErrors();
   });
 
+  it('Should be able to set id/automation id single', async () => {
+    await browser.driver.sleep(config.sleep);
+
+    expect(await element(by.id('bullet-example1-title')).getAttribute('id')).toEqual('bullet-example1-title');
+    expect(await element(by.id('bullet-example1-title')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-example1-title');
+
+    expect(await element(by.id('bullet-example1-subtitle')).getAttribute('id')).toEqual('bullet-example1-subtitle');
+    expect(await element(by.id('bullet-example1-subtitle')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-example1-subtitle');
+
+    expect(await element(by.id('bullet-example1-difference')).getAttribute('id')).toEqual('bullet-example1-difference');
+    expect(await element(by.id('bullet-example1-difference')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-example1-difference');
+
+    expect(await element(by.id('bullet-example1-marker')).getAttribute('id')).toEqual('bullet-example1-marker');
+    expect(await element(by.id('bullet-example1-marker')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-example1-marker');
+
+    expect(await element(by.id('bullet-example1-measure0')).getAttribute('id')).toEqual('bullet-example1-measure0');
+    expect(await element(by.id('bullet-example1-measure0')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-example1-measure0');
+    expect(await element(by.id('bullet-example1-measure1')).getAttribute('id')).toEqual('bullet-example1-measure1');
+    expect(await element(by.id('bullet-example1-measure1')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-example1-measure1');
+
+    expect(await element(by.id('bullet-example1-range0')).getAttribute('id')).toEqual('bullet-example1-range0');
+    expect(await element(by.id('bullet-example1-range0')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-example1-range0');
+    expect(await element(by.id('bullet-example1-range1')).getAttribute('id')).toEqual('bullet-example1-range1');
+    expect(await element(by.id('bullet-example1-range1')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-example1-range1');
+    expect(await element(by.id('bullet-example1-range2')).getAttribute('id')).toEqual('bullet-example1-range2');
+    expect(await element(by.id('bullet-example1-range2')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-example1-range2');
+    expect(await element(by.id('bullet-example1-range3')).getAttribute('id')).toEqual('bullet-example1-range3');
+    expect(await element(by.id('bullet-example1-range3')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-example1-range3');
+    expect(await element(by.id('bullet-example1-range4')).getAttribute('id')).toEqual('bullet-example1-range4');
+    expect(await element(by.id('bullet-example1-range4')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-example1-range4');
+  });
+
   if (utils.isChrome() && utils.isCI()) {
     it('Should not visual regress', async () => {
       const containerEl = await element(by.css('div[role=main]'));
@@ -47,6 +79,116 @@ describe('Bullet data group tests', () => {
       expect(await browser.imageComparison.checkScreen('bullet-data-group')).toEqual(0);
     });
   }
+});
+
+describe('Bullet data group automation tests', () => {
+  beforeEach(async () => {
+    await utils.setPage('/components/bullet/test-data-group-automation?layout=nofrills');
+  });
+
+  it('Should not have errors', async () => {
+    await utils.checkForErrors();
+  });
+
+  it('Should be able to set id/automation id grouped', async () => {
+    await browser.driver.sleep(config.sleep);
+
+    expect(await element(by.id('bullet-group-example1-title-group0')).getAttribute('id')).toEqual('bullet-group-example1-title-group0');
+    expect(await element(by.id('bullet-group-example1-title-group0')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-title-group0');
+    expect(await element(by.id('bullet-group-example1-title-group1')).getAttribute('id')).toEqual('bullet-group-example1-title-group1');
+    expect(await element(by.id('bullet-group-example1-title-group1')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-title-group1');
+    expect(await element(by.id('bullet-group-example1-title-group2')).getAttribute('id')).toEqual('bullet-group-example1-title-group2');
+    expect(await element(by.id('bullet-group-example1-title-group2')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-title-group2');
+    expect(await element(by.id('bullet-group-example1-title-group3')).getAttribute('id')).toEqual('bullet-group-example1-title-group3');
+    expect(await element(by.id('bullet-group-example1-title-group3')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-title-group3');
+    expect(await element(by.id('bullet-group-example1-title-group4')).getAttribute('id')).toEqual('bullet-group-example1-title-group4');
+    expect(await element(by.id('bullet-group-example1-title-group4')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-title-group4');
+
+    expect(await element(by.id('bullet-group-example1-subtitle-group0')).getAttribute('id')).toEqual('bullet-group-example1-subtitle-group0');
+    expect(await element(by.id('bullet-group-example1-subtitle-group0')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-subtitle-group0');
+    expect(await element(by.id('bullet-group-example1-subtitle-group1')).getAttribute('id')).toEqual('bullet-group-example1-subtitle-group1');
+    expect(await element(by.id('bullet-group-example1-subtitle-group1')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-subtitle-group1');
+    expect(await element(by.id('bullet-group-example1-subtitle-group2')).getAttribute('id')).toEqual('bullet-group-example1-subtitle-group2');
+    expect(await element(by.id('bullet-group-example1-subtitle-group2')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-subtitle-group2');
+    expect(await element(by.id('bullet-group-example1-subtitle-group3')).getAttribute('id')).toEqual('bullet-group-example1-subtitle-group3');
+    expect(await element(by.id('bullet-group-example1-subtitle-group3')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-subtitle-group3');
+    expect(await element(by.id('bullet-group-example1-subtitle-group4')).getAttribute('id')).toEqual('bullet-group-example1-subtitle-group4');
+    expect(await element(by.id('bullet-group-example1-subtitle-group4')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-subtitle-group4');
+
+    expect(await element(by.id('bullet-group-example1-difference-group0')).getAttribute('id')).toEqual('bullet-group-example1-difference-group0');
+    expect(await element(by.id('bullet-group-example1-difference-group0')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-difference-group0');
+    expect(await element(by.id('bullet-group-example1-difference-group1')).getAttribute('id')).toEqual('bullet-group-example1-difference-group1');
+    expect(await element(by.id('bullet-group-example1-difference-group1')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-difference-group1');
+    expect(await element(by.id('bullet-group-example1-difference-group2')).getAttribute('id')).toEqual('bullet-group-example1-difference-group2');
+    expect(await element(by.id('bullet-group-example1-difference-group2')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-difference-group2');
+    expect(await element(by.id('bullet-group-example1-difference-group3')).getAttribute('id')).toEqual('bullet-group-example1-difference-group3');
+    expect(await element(by.id('bullet-group-example1-difference-group3')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-difference-group3');
+    expect(await element(by.id('bullet-group-example1-difference-group4')).getAttribute('id')).toEqual('bullet-group-example1-difference-group4');
+    expect(await element(by.id('bullet-group-example1-difference-group4')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-difference-group4');
+
+    expect(await element(by.id('bullet-group-example1-marker-group0')).getAttribute('id')).toEqual('bullet-group-example1-marker-group0');
+    expect(await element(by.id('bullet-group-example1-marker-group0')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-marker-group0');
+    expect(await element(by.id('bullet-group-example1-marker-group1')).getAttribute('id')).toEqual('bullet-group-example1-marker-group1');
+    expect(await element(by.id('bullet-group-example1-marker-group1')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-marker-group1');
+    expect(await element(by.id('bullet-group-example1-marker-group2')).getAttribute('id')).toEqual('bullet-group-example1-marker-group2');
+    expect(await element(by.id('bullet-group-example1-marker-group2')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-marker-group2');
+    expect(await element(by.id('bullet-group-example1-marker-group3')).getAttribute('id')).toEqual('bullet-group-example1-marker-group3');
+    expect(await element(by.id('bullet-group-example1-marker-group3')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-marker-group3');
+    expect(await element(by.id('bullet-group-example1-marker-group4')).getAttribute('id')).toEqual('bullet-group-example1-marker-group4');
+    expect(await element(by.id('bullet-group-example1-marker-group4')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-marker-group4');
+
+    expect(await element(by.id('bullet-group-example1-measure0-group0')).getAttribute('id')).toEqual('bullet-group-example1-measure0-group0');
+    expect(await element(by.id('bullet-group-example1-measure0-group0')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-measure0-group0');
+    expect(await element(by.id('bullet-group-example1-measure1-group0')).getAttribute('id')).toEqual('bullet-group-example1-measure1-group0');
+    expect(await element(by.id('bullet-group-example1-measure1-group0')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-measure1-group0');
+    expect(await element(by.id('bullet-group-example1-measure0-group1')).getAttribute('id')).toEqual('bullet-group-example1-measure0-group1');
+    expect(await element(by.id('bullet-group-example1-measure0-group1')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-measure0-group1');
+    expect(await element(by.id('bullet-group-example1-measure1-group1')).getAttribute('id')).toEqual('bullet-group-example1-measure1-group1');
+    expect(await element(by.id('bullet-group-example1-measure1-group1')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-measure1-group1');
+    expect(await element(by.id('bullet-group-example1-measure0-group2')).getAttribute('id')).toEqual('bullet-group-example1-measure0-group2');
+    expect(await element(by.id('bullet-group-example1-measure0-group2')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-measure0-group2');
+    expect(await element(by.id('bullet-group-example1-measure1-group2')).getAttribute('id')).toEqual('bullet-group-example1-measure1-group2');
+    expect(await element(by.id('bullet-group-example1-measure1-group2')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-measure1-group2');
+    expect(await element(by.id('bullet-group-example1-measure0-group3')).getAttribute('id')).toEqual('bullet-group-example1-measure0-group3');
+    expect(await element(by.id('bullet-group-example1-measure0-group3')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-measure0-group3');
+    expect(await element(by.id('bullet-group-example1-measure1-group3')).getAttribute('id')).toEqual('bullet-group-example1-measure1-group3');
+    expect(await element(by.id('bullet-group-example1-measure1-group3')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-measure1-group3');
+    expect(await element(by.id('bullet-group-example1-measure0-group4')).getAttribute('id')).toEqual('bullet-group-example1-measure0-group4');
+    expect(await element(by.id('bullet-group-example1-measure0-group4')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-measure0-group4');
+    expect(await element(by.id('bullet-group-example1-measure1-group4')).getAttribute('id')).toEqual('bullet-group-example1-measure1-group4');
+    expect(await element(by.id('bullet-group-example1-measure1-group4')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-measure1-group4');
+
+    expect(await element(by.id('bullet-group-example1-range0-group0')).getAttribute('id')).toEqual('bullet-group-example1-range0-group0');
+    expect(await element(by.id('bullet-group-example1-range0-group0')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range0-group0');
+    expect(await element(by.id('bullet-group-example1-range1-group0')).getAttribute('id')).toEqual('bullet-group-example1-range1-group0');
+    expect(await element(by.id('bullet-group-example1-range1-group0')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range1-group0');
+    expect(await element(by.id('bullet-group-example1-range2-group0')).getAttribute('id')).toEqual('bullet-group-example1-range2-group0');
+    expect(await element(by.id('bullet-group-example1-range2-group0')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range2-group0');
+    expect(await element(by.id('bullet-group-example1-range0-group1')).getAttribute('id')).toEqual('bullet-group-example1-range0-group1');
+    expect(await element(by.id('bullet-group-example1-range0-group1')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range0-group1');
+    expect(await element(by.id('bullet-group-example1-range1-group1')).getAttribute('id')).toEqual('bullet-group-example1-range1-group1');
+    expect(await element(by.id('bullet-group-example1-range1-group1')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range1-group1');
+    expect(await element(by.id('bullet-group-example1-range2-group1')).getAttribute('id')).toEqual('bullet-group-example1-range2-group1');
+    expect(await element(by.id('bullet-group-example1-range2-group1')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range2-group1');
+    expect(await element(by.id('bullet-group-example1-range0-group2')).getAttribute('id')).toEqual('bullet-group-example1-range0-group2');
+    expect(await element(by.id('bullet-group-example1-range0-group2')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range0-group2');
+    expect(await element(by.id('bullet-group-example1-range1-group2')).getAttribute('id')).toEqual('bullet-group-example1-range1-group2');
+    expect(await element(by.id('bullet-group-example1-range1-group2')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range1-group2');
+    expect(await element(by.id('bullet-group-example1-range2-group2')).getAttribute('id')).toEqual('bullet-group-example1-range2-group2');
+    expect(await element(by.id('bullet-group-example1-range2-group2')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range2-group2');
+    expect(await element(by.id('bullet-group-example1-range0-group3')).getAttribute('id')).toEqual('bullet-group-example1-range0-group3');
+    expect(await element(by.id('bullet-group-example1-range0-group3')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range0-group3');
+    expect(await element(by.id('bullet-group-example1-range1-group3')).getAttribute('id')).toEqual('bullet-group-example1-range1-group3');
+    expect(await element(by.id('bullet-group-example1-range1-group3')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range1-group3');
+    expect(await element(by.id('bullet-group-example1-range2-group3')).getAttribute('id')).toEqual('bullet-group-example1-range2-group3');
+    expect(await element(by.id('bullet-group-example1-range2-group3')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range2-group3');
+    expect(await element(by.id('bullet-group-example1-range0-group4')).getAttribute('id')).toEqual('bullet-group-example1-range0-group4');
+    expect(await element(by.id('bullet-group-example1-range0-group4')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range0-group4');
+    expect(await element(by.id('bullet-group-example1-range1-group4')).getAttribute('id')).toEqual('bullet-group-example1-range1-group4');
+    expect(await element(by.id('bullet-group-example1-range1-group4')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range1-group4');
+    expect(await element(by.id('bullet-group-example1-range2-group4')).getAttribute('id')).toEqual('bullet-group-example1-range2-group4');
+    expect(await element(by.id('bullet-group-example1-range2-group4')).getAttribute('data-automation-id')).toEqual('automation-id-bullet-group-example1-range2-group4');
+  });
 });
 
 describe('Bullet negative positive tests', () => {